ARMADA-460 refactoring get_s3_object_as_tarfile
It was moved to s3.manager.py to be accessed from multiple endpoints
