+++ jobbergate-api/jobbergate_api/tests/apps/job_scripts/test_routers.py
@@ -9,13 +9,11 @@ import asyncpg
 import pytest
 from botocore.exceptions import BotoCoreError
 from fastapi import status
-from fastapi.exceptions import HTTPException
 
 from jobbergate_api.apps.applications.models import applications_table
 from jobbergate_api.apps.job_scripts.models import job_scripts_table
 from jobbergate_api.apps.job_scripts.routers import (
     build_job_script_data_as_string,
-    get_s3_object_as_tarfile,
     inject_sbatch_params,
     render_template,
     s3man,
@@ -251,38 +249,6 @@ async def test_create_job_script_file_not_found(
 
     count = await database.fetch_all("SELECT COUNT(*) FROM job_scripts")
     assert count[0][0] == 0
-
-
-@pytest.mark.asyncio
-@mock.patch.object(s3man.client, "s3_client")
-@database.transaction(force_rollback=True)
-async def test_get_s3_object_as_tarfile(s3man_client_mock, s3_object):
-    """
-    Test getting a file from S3 with get_s3_object function.
-    """
-    s3man_client_mock.get_object.return_value = s3_object
-
-    s3_file = get_s3_object_as_tarfile(1)
-
-    assert s3_file is not None
-    s3man_client_mock.get_object.assert_called_once()
-
-
-@mock.patch.object(s3man.client, "s3_client")
-def test_get_s3_object_not_found(s3man_client_mock):
-    """
-    Test exception when file not exists in S3 for get_s3_object function.
-    """
-    s3man_client_mock.get_object.side_effect = BotoCoreError()
-
-    s3_file = None
-    with pytest.raises(HTTPException) as exc:
-        s3_file = get_s3_object_as_tarfile(1)
-
-    assert "Application with id=1 not found" in str(exc)
-
-    assert s3_file is None
-    s3man_client_mock.get_object.assert_called_once()
 
 
 def test_render_template(param_dict_flat, template_files, job_script_data_as_string):
